Move zoom function to utils
